Refs #24682: Storyblok - replace local `BaseImage` component with a global one
